A1779617_KaifengChen/preProcessing/requirements.txt
gunicorn==20.1.0
Flask==2.0.2
google-cloud-storage
opencv-python
imutils
werkzeug==2.0.2